requirements.txt
boto3>=1.26.0
PyYAML>=6.0
psycopg2-binary>=2.9.0
click>=8.0.0
tqdm>=4.64.0

# Optional cloud storage dependencies
google-cloud-storage>=2.5.0; extra == 'gcs'
azure-storage-blob>=12.0.0; extra == 'azure'